questione id e nome agenzia da sistemare

+++ database/migrations/2022_02_06_142247_create_employments_table.php
@@ -18,7 +18,7 @@ class CreateEmploymentsTable extends Migration
             $table->string('nome');
             $table->string('cognome');
             $table->string('azienda');
-            $table->unsignedSmallInteger('cell');
+            $table->unsignedInteger('cell');
             $table->string('mail')->unique();
             $table->timestamps();
         });
